Made letter folder in modules and put all letter related files in there
