+++ src/admin/redux/modules/letters/letters.js
@@ -11,7 +11,8 @@ const {
   , liftN
   , unapply
 } = require('ramda')
-const formatDate = require('app/format_date')
+
+const { sub_reminder } = require('./bodies.js')
 
 const SEND_NEWSLETTER_POST =
   'SEND_NEWSLETTER_POST'
@@ -34,7 +35,7 @@ const reducer: Reducer<State, Action>
      return { ...state, post_members: payload.results }
    case SEND_SUB_REMINDER_POST:
      const ids = pick([ 'id' ])
-     const emails = compose(objOf('email_content'), inject)
+     const emails = compose(objOf('email_content'), sub_reminder)
      const shape = map(liftN(3, unapply(reduce(merge, {})))(emails, addresses, ids))
      return { ...state, sub_reminders: shape(payload.results) }
    default:
@@ -52,20 +53,3 @@ export const send_sub_reminder_post =
 
 const addressProps = [ 'address1', 'address2', 'address3', 'address4', 'county', 'postcode' ]
 const addresses = compose(objOf('address'), props(addressProps))
-
-const getOverdue = (days) => {
-  if (days > 90) return 90
-  return days > 60 ? 60 : 30
-}
-
-const inject = (members) => (
-  `Dear ${members.first_name || members.title + ' ' + members.last_name },
-  We notice that your Standing Order which is normally paid on ${formatDate(members.due_date)}
-  each year has not been paid this year and £${members.amount} has now been unpaid
-  for over ${getOverdue(members.overdue)} days. We assume that this is probably an
-  administrative error and would be very grateful if you could look into it.
-  If, alternatively, your intention is to cancel your membership of the Friends
-  we’d be grateful if you could let the Membership Secretary
-  (Pam Marrs, 42 Bracklesham Road, Hayling Island PO11 9SJ) know that that is your intention.
-  If you have already sorted the problem out, our apologies and please ignore this letter.`
-)
